Implement mod creation

+++ src-tauri/src/commands.rs
@@ -108,15 +108,6 @@ pub fn get_string_tables(app: tauri::AppHandle) -> StringTables {
 #[tauri::command]
 pub fn get_mods(app: tauri::AppHandle) -> Vec<String> {
     get_mod_names(&app)
-    /*vec![
-        "a".to_string(),
-        "b".to_string(),
-        "c".to_string(),
-        "d".to_string(),
-        "e".to_string(),
-        "f".to_string(),
-        "g".to_string(),
-    ]*/
 }
 
 #[tauri::command]
@@ -127,10 +118,6 @@ pub fn save_mod(app: tauri::AppHandle, mod_name: &str) {
     for file in MOD_FILES.iter() {
         let source = temp_directory.join(file);
         let destination = mod_directory.join(file);
-
-        println!("---------");
-        println!("source: {source:?}");
-        println!("dest  : {destination:?}");
 
         fs::copy(source, destination).unwrap();
     }
@@ -145,10 +132,16 @@ pub fn load_mod(app: tauri::AppHandle, mod_name: &str) {
         let source = mod_directory.join(file);
         let destination = temp_directory.join(file);
 
-        println!("---------");
-        println!("source: {source:?}");
-        println!("dest  : {destination:?}");
-
-        fs::copy(source, destination).unwrap();
+        if source.exists() {
+            fs::copy(source, destination).unwrap();
+        }
     }
+}
+
+#[tauri::command]
+pub fn create_mod(app: tauri::AppHandle, mod_name: &str) {
+    let mod_directory = get_mod_directory(&app, mod_name);
+
+    fs::remove_dir_all(mod_directory).unwrap();
+    get_mod_directory(&app, mod_name);
 }
